Move storage checks to a mod
