+++ chainstate/launcher/src/lib.rs
@@ -16,13 +16,12 @@
 //! Tools to set up chainstate together with its storage
 
 mod config;
+mod storage_compatibility;
 
 use std::sync::Arc;
 
 use chainstate::InitializationError;
-use chainstate_storage::{BlockchainStorageRead, BlockchainStorageWrite};
 use storage_lmdb::resize_callback::MapResizeCallback;
-use utils::ensure;
 
 // Some useful reexports
 pub use chainstate::{
@@ -35,10 +34,6 @@ pub use config::{ChainstateLauncherConfig, StorageBackendConfig};
 /// Subdirectory under `datadir` where LMDB chainstate database is placed
 pub const SUBDIRECTORY_LMDB: &str = "chainstate-lmdb";
 
-const CHAINSTATE_STORAGE_VERSION_UNINITIALIZED: u32 = 0;
-const CHAINSTATE_STORAGE_VERSION_V1: u32 = 1;
-const CURRENT_CHAINSTATE_STORAGE_VERSION: u32 = CHAINSTATE_STORAGE_VERSION_V1;
-
 fn make_chainstate_and_storage_impl<B: 'static + storage::Backend>(
     storage_backend: B,
     chain_config: Arc<ChainConfig>,
@@ -47,9 +42,8 @@ fn make_chainstate_and_storage_impl<B: 'static + storage::Backend>(
     let mut storage = chainstate_storage::Store::new(storage_backend)
         .map_err(|e| Error::FailedToInitializeChainstate(e.into()))?;
 
-    check_storage_version(&mut storage)?;
-    check_magic_bytes(&mut storage, chain_config.as_ref())?;
-    check_chain_type(&mut storage, chain_config.as_ref())?;
+    storage_compatibility::check_storage_compatibility(&mut storage, chain_config.as_ref())
+        .map_err(InitializationError::StorageCompatibilityCheckError)?;
 
     let chainstate = chainstate::make_chainstate(
         chain_config,
@@ -60,75 +54,6 @@ fn make_chainstate_and_storage_impl<B: 'static + storage::Backend>(
         Default::default(),
     )?;
     Ok(chainstate)
-}
-
-fn check_storage_version<B: 'static + storage::Backend>(
-    storage: &mut chainstate_storage::Store<B>,
-) -> Result<(), Error> {
-    let storage_version =
-        storage.get_storage_version().map_err(InitializationError::StorageError)?;
-
-    if storage_version == CHAINSTATE_STORAGE_VERSION_UNINITIALIZED {
-        storage
-            .set_storage_version(CURRENT_CHAINSTATE_STORAGE_VERSION)
-            .map_err(InitializationError::StorageError)?;
-    } else {
-        ensure!(
-            storage_version == CURRENT_CHAINSTATE_STORAGE_VERSION,
-            InitializationError::ChainstateStorageVersionMismatch(
-                storage_version,
-                CURRENT_CHAINSTATE_STORAGE_VERSION
-            )
-        );
-    }
-    Ok(())
-}
-
-fn check_magic_bytes<B: 'static + storage::Backend>(
-    storage: &mut chainstate_storage::Store<B>,
-    chain_config: &ChainConfig,
-) -> Result<(), Error> {
-    let storage_magic_bytes =
-        storage.get_magic_bytes().map_err(InitializationError::StorageError)?;
-    let chain_config_magic_bytes = chain_config.magic_bytes();
-
-    match storage_magic_bytes {
-        Some(storage_magic_bytes) => ensure!(
-            &storage_magic_bytes == chain_config_magic_bytes,
-            InitializationError::ChainConfigMagicBytesMismatch(
-                storage_magic_bytes,
-                chain_config_magic_bytes.to_owned()
-            )
-        ),
-        None => storage
-            .set_magic_bytes(chain_config_magic_bytes)
-            .map_err(InitializationError::StorageError)?,
-    };
-
-    Ok(())
-}
-
-fn check_chain_type<B: 'static + storage::Backend>(
-    storage: &mut chainstate_storage::Store<B>,
-    chain_config: &ChainConfig,
-) -> Result<(), Error> {
-    let storage_chain_type = storage.get_chain_type().map_err(InitializationError::StorageError)?;
-    let chain_config_type = chain_config.chain_type().name();
-
-    match storage_chain_type {
-        Some(storage_chain_type) => ensure!(
-            storage_chain_type == chain_config_type,
-            InitializationError::ChainTypeMismatch(
-                storage_chain_type,
-                chain_config_type.to_owned()
-            )
-        ),
-        None => storage
-            .set_chain_type(chain_config_type)
-            .map_err(InitializationError::StorageError)?,
-    };
-
-    Ok(())
 }
 
 /// Create chainstate together with its storage
